Convert json into string array
