+++ src/app/decision-tree/family-questions-store.ts
@@ -1,0 +1,40 @@
+import { create } from "zustand";
+
+type FamilyQuestions = {
+  questions: [];
+  addQuestion: (question: string) => void;
+};
+
+interface FamilyQuestionsState {
+  questions: string[];
+    addQuestion: (question: string) => void;
+}
+
+export const useFamilyQuestionsStore = create<FamilyQuestionsState>((set) => ({
+  questions: [
+    "Does this align with our family’s values?",
+    "Is this our greatest priority, or in the top 3 at the moment?",
+    "Am I the best person to be managing this decision?",
+    "Is this something impactful for our family or the world?",
+    "Is this the best use of my family’s time, energy, and capital?",
+    "Why is this something that I, or another family member is passionate about or feels is important?",
+    "Why do I trust this person/investment/charity? Have I known them for a year?",
+    "Do others trust them?",
+    "If this project/investment/endeavour fails, can we/I withstand the loss? And if not, is it more important than the risk? Success or failure, what have I learned?",
+    "How have I verified this information being presented to me?",
+    "Is there a way I can trial or test this investment/person/opportunity before we fully commit?",
+    "Have I consulted at least one other family member and one of our family’s trusted advisors on this decision?",
+    "In 4 generations will this be important to my family and the world?"
+],
+  addQuestion: (question: string) => set((state) => ({ questions: [...state.questions, question] })),
+}));
+
+export const useFamilyQuestions = () => {
+  return {
+    questions: useFamilyQuestionsStore((s) => s.questions),
+    addQuestion: useFamilyQuestionsStore((s) => s.addQuestion),
+  };
+};
+
+
+
