+++ src/app/page.tsx
@@ -1,18 +1,20 @@
 import "./globals.css";
-import AddQuestion from "../app/components/AddQuestion";
-import Questions from "../app/components/Questions";
+import AddQuestion from "./decision-tree/AddQuestion";
+import Questions from "./decision-tree/Questions";
 import { IQuestion } from "../../types/questions";
+import Link from "next/link";
 
 export default async function Home() {
   const questions = [] as IQuestion[];
   console.log(questions);
   return (
     <main className="max-w-4xl mx-auto mt-4">
-      <div className="text-center text-4xl  my-5 flex flex-col gap-4 ">
-        <h1 className="text-5xl font-bold">Decision Tree</h1>
-        <AddQuestion />
-      </div>
-      <Questions questions={questions} />
+      <Link href="/decision-tree">Decision Tree</Link>
+      <Link href=""></Link>
+      {/* <Link></Link>
+      <Link></Link>
+      <Link></Link>
+      <Link></Link> */}
     </main>
   );
 }
